thycotic-secret/requirements.txt
PyYaml
zeep
